+++ Gw2Sharp.Tests/ConnectionTests.cs
@@ -71,6 +71,24 @@ namespace Gw2Sharp.Tests
         }
 
         [Theory]
+        [InlineData("55BCD46F-B57C-469A-AE81-3B21B5583573", true)]
+        [InlineData("eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c", true)]
+        [InlineData("\"", false)]
+        public void ValidAccessTokenTest(string accessToken, bool isValid)
+        {
+            if (!isValid)
+                Assert.Throws<ArgumentException>("accessToken", () => new Connection(accessToken));
+            else
+                new Connection(accessToken); // Should not throw
+
+            var connection = new Connection();
+            if (!isValid)
+                Assert.Throws<ArgumentException>("value", () => connection.AccessToken = accessToken);
+            else
+                connection.AccessToken = accessToken; // Should not throw
+        }
+
+        [Theory]
         [InlineData("en", Locale.English)]
         [InlineData("de", Locale.German)]
         [InlineData("fr", Locale.French)]
@@ -181,8 +199,8 @@ namespace Gw2Sharp.Tests
         {
             var connection = new Connection();
             await AssertArguments.ThrowsWhenNullAsync(
-                 () => connection.RequestAsync<object>(null, new Uri("http://localhost"), null, CancellationToken.None),
-                 new[] { false, true, false, false });
+                () => connection.RequestAsync<object>(null, new Uri("http://localhost"), null, CancellationToken.None),
+                new[] { false, true, false, false });
         }
 
         #endregion